requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings
python-multipart
pydantic[email]